More refactor, remove synapse_version as api
