+++ src/charm.py
@@ -36,6 +36,7 @@ logger = logging.getLogger(__name__)
 
 INGRESS_INTEGRATION_NAME = "ingress"
 MAIN_UNIT_ID = 0
+SYNAPSE_VERSION = "1.132.0"
 
 
 class SynapseCharm(CharmBaseWithState):
@@ -57,6 +58,7 @@ class SynapseCharm(CharmBaseWithState):
             args: class arguments.
         """
         super().__init__(*args)
+        self.unit.set_workload_version(SYNAPSE_VERSION)
         self._backup = BackupObserver(self)
         self._matrix_auth = MatrixAuthObserver(self)
         self._media = MediaObserver(self)
@@ -111,7 +113,7 @@ class SynapseCharm(CharmBaseWithState):
             media_config=self._media.get_relation_as_media_conf(),
             redis_config=self._redis.get_relation_as_redis_conf(),
             registration_secrets=self._matrix_auth.get_requirer_registration_secrets(),
-            instance_map_config=self.instance_map(),
+            instance_map_config=self._instance_map(),
         )
 
     @inject_charm_state
@@ -165,59 +167,38 @@ class SynapseCharm(CharmBaseWithState):
         if not container.can_connect():
             self.unit.status = ops.MaintenanceStatus("Waiting for Synapse pebble")
             return
-        if charm_state.redis_config is None and self.app.planned_units() > 1:
-            logger.debug("More than 1 peer unit found. Redis is required.")
+        if self._is_redis_required(charm_state):
             self.unit.status = ops.BlockedStatus("Redis integration is required.")
             return
         self.model.unit.status = ops.MaintenanceStatus("Configuring Synapse")
         try:
-            # check signing key
-            signing_key_path = f"/data/{charm_state.synapse_config.server_name}.signing.key"
-            signing_key_from_secret = self.get_signing_key()
-            if signing_key_from_secret:
-                logger.debug("Signing key secret was found, pushing it to the container")
-                container.push(
-                    signing_key_path, signing_key_from_secret, make_dirs=True, encoding="utf-8"
-                )
-
-            # reconcile configuration
-            pebble.reconcile(
-                charm_state, container, is_main=self.is_main(), unit_number=self.get_unit_number()
-            )
-
-            # create new signing key if needed
-            if self.is_main() and not signing_key_from_secret:
-                logger.debug("Signing key secret not found, creating secret")
-                with container.pull(signing_key_path) as f:
-                    signing_key = f.read()
-                    self.set_signing_key(signing_key.rstrip())
-
-            # update matrix-auth integration with configuration data
-            if self.unit.is_leader():
-                self._matrix_auth.update_matrix_auth_integration(charm_state)
-
-            # set app status with workload version
-            synapse_version = synapse.get_version(self.get_unit_address(MAIN_UNIT_ID))
-            self.unit.set_workload_version(synapse_version)
-
-            # enable Mjolnir
-            if self.is_main():
-                self._mjolnir.enable_mjolnir(charm_state)
-        except synapse.APIError as exc:
-            logger.debug("Cannot set workload version at this time: %s", exc)
+            self._configure_and_start_services(charm_state, container)
         except (pebble.PebbleServiceError, FileNotFoundError) as exc:
             self.model.unit.status = ops.BlockedStatus(str(exc))
             return
-        pebble.restart_nginx(container, self.get_unit_address(MAIN_UNIT_ID))
+        if self.unit.is_leader():
+            self._matrix_auth.update_matrix_auth_integration(charm_state)
         self._set_unit_status()
 
-    def peers(self) -> typing.Optional[ops.Relation]:
-        """Get peer relation.
+    def manage_signing_key(self, charm_state: CharmState, container: ops.Container) -> None:
+        """Handle signing key secret pull and push.
 
-        Returns:
-            Synapse peer relation.
+        Args:
+            charm_state: charm state.
+            container: container.
         """
-        return self.model.get_relation(synapse.SYNAPSE_PEER_RELATION_NAME)
+        signing_key_path = f"/data/{charm_state.synapse_config.server_name}.signing.key"
+        signing_key_from_secret = self.get_signing_key()
+        if signing_key_from_secret:
+            logger.debug("Signing key secret was found, pushing it to the container")
+            container.push(
+                signing_key_path, signing_key_from_secret, make_dirs=True, encoding="utf-8"
+            )
+        if self._is_main() and not signing_key_from_secret:
+            logger.debug("Signing key secret not found, creating secret")
+            with container.pull(signing_key_path) as f:
+                signing_key = f.read()
+                self.set_signing_key(signing_key.rstrip())
 
     def get_signing_key(self) -> typing.Optional[str]:
         """Get signing key from secret.
@@ -225,7 +206,7 @@ class SynapseCharm(CharmBaseWithState):
         Returns:
             Signing key as string or None if not found.
         """
-        peer_relation = self.peers()
+        peer_relation = self._peers()
         if not peer_relation:
             logger.error(
                 "Failed to get signing key: no peer relation %s found",
@@ -243,40 +224,13 @@ class SynapseCharm(CharmBaseWithState):
                 del peer_relation.data[self.app]["secret-signing-id"]
         return None
 
-    def is_main(self) -> bool:
-        """Verify if this unit is the main.
-
-        Returns:
-            bool: true if is the main unit.
-        """
-        return f"/{MAIN_UNIT_ID}" in self.unit.name
-
-    def get_unit_address(self, unit_id: int) -> str:
-        """Get unit address.
-
-        Args:
-            unit_id: number as 0 in synapse/0.
-
-        Returns:
-            unit address as unit-0.synapse-endpoints.
-        """
-        return f"{self.app.name}-{unit_id}.{self.app.name}-endpoints"
-
-    def get_unit_number(self) -> str:
-        """Get unit number.
-
-        Returns:
-            unit number as 0 in synapse/0.
-        """
-        return self.unit.name.split("/")[1]
-
     def set_signing_key(self, signing_key: str) -> None:
         """Create secret with signing key content.
 
         Args:
             signing_key: signing key as string.
         """
-        peer_relation = self.peers()
+        peer_relation = self._peers()
         if not peer_relation:
             logger.error(
                 "Failed to set signing key: no peer relation %s found",
@@ -292,13 +246,68 @@ class SynapseCharm(CharmBaseWithState):
             secret = self.app.add_secret({"secret-signing-key": signing_key})
             peer_relation.data[self.app].update({"secret-signing-id": typing.cast(str, secret.id)})
 
-    def instance_map(self) -> typing.Optional[typing.Dict]:
+    def _configure_and_start_services(
+        self, charm_state: CharmState, container: ops.Container
+    ) -> None:
+        """Configure and start pebble layers."""
+        self.manage_signing_key(charm_state, container)
+        pebble.reconcile(
+            charm_state, container, is_main=self._is_main(), unit_number=self._get_unit_number()
+        )
+        if self._is_main() and charm_state.synapse_config.enable_mjolnir:
+            self._mjolnir.enable(charm_state)
+        pebble.restart_nginx(container, self._get_unit_address(MAIN_UNIT_ID))
+
+    def _is_redis_required(self, charm_state: CharmState) -> bool:
+        """Check if Redis configuration should be required.
+
+        Return:
+            True if more than 1 unit is found.
+        """
+        return charm_state.redis_config is None and self.app.planned_units() > 1
+
+    def _peers(self) -> typing.Optional[ops.Relation]:
+        """Get peer relation.
+
+        Returns:
+            Synapse peer relation.
+        """
+        return self.model.get_relation(synapse.SYNAPSE_PEER_RELATION_NAME)
+
+    def _is_main(self) -> bool:
+        """Verify if this unit is the main.
+
+        Returns:
+            bool: true if is the main unit.
+        """
+        return f"/{MAIN_UNIT_ID}" in self.unit.name
+
+    def _get_unit_address(self, unit_id: int) -> str:
+        """Get unit address.
+
+        Args:
+            unit_id: number as 0 in synapse/0.
+
+        Returns:
+            unit address as unit-0.synapse-endpoints.
+        """
+        return f"{self.app.name}-{unit_id}.{self.app.name}-endpoints"
+
+    def _get_unit_number(self) -> str:
+        """Get unit number.
+
+        Returns:
+            unit number as 0 in synapse/0.
+        """
+        return self.unit.name.split("/")[1]
+
+    def _instance_map(self) -> typing.Optional[typing.Dict]:
         """Create instance_map configuration.
 
         Returns:
             Instance map configuration as a dict or None if there is only one unit.
         """
-        peer_relation = self.peers()
+        peer_relation = self._peers()
         if not peer_relation or len(peer_relation.units) <= 1:
             logger.debug("One unit in peer relation, skipping instance_map configuration")
             return None
@@ -306,17 +315,17 @@ class SynapseCharm(CharmBaseWithState):
         for unit_id in range(len(peer_relation.units)):
             if unit_id == MAIN_UNIT_ID:
                 instance_map["main"] = {
-                    "host": self.get_unit_address(MAIN_UNIT_ID),
+                    "host": self._get_unit_address(MAIN_UNIT_ID),
                     "port": 8035,
                 }
                 instance_map["federationsender1"] = {
-                    "host": self.get_unit_address(MAIN_UNIT_ID),
+                    "host": self._get_unit_address(MAIN_UNIT_ID),
                     "port": 8034,
                 }
                 continue
 
             instance_name = f"worker{unit_id}"
-            address = self.get_unit_address(unit_id)
+            address = self._get_unit_address(unit_id)
             instance_map[instance_name] = {
                 "host": address,
                 "port": 8034,
@@ -325,37 +334,33 @@ class SynapseCharm(CharmBaseWithState):
 
     def _set_unit_status(self) -> None:
         """Set unit status depending on Synapse and NGINX state."""
-        # This method contains a similar check that the one in mjolnir.py for Synapse
-        # container and service. Until a refactoring is done for a different way of checking
-        # and setting the unit status in a hollistic way, both checks will be in place.
-        # pylint: disable=R0801
-
-        # If the unit is in a blocked state, do not change it, as it
-        # was set by a problem or error with the configuration
         if isinstance(self.unit.status, ops.BlockedStatus):
+            # Preserve error state set elsewhere
             return
-        # Synapse checks
+
         container = self.unit.get_container(synapse.SYNAPSE_CONTAINER_NAME)
         if not container.can_connect():
             self.unit.status = ops.MaintenanceStatus("Waiting for Synapse pebble")
             return
-        synapse_service = container.get_services(synapse.SYNAPSE_SERVICE_NAME)
-        synapse_not_active = [
-            service for service in synapse_service.values() if not service.is_running()
-        ]
-        if not synapse_service or synapse_not_active:
+
+        if not self._is_service_running(container, synapse.SYNAPSE_SERVICE_NAME):
             self.unit.status = ops.MaintenanceStatus("Waiting for Synapse")
             return
-        # NGINX checks
-        nginx_service = container.get_services(synapse.SYNAPSE_NGINX_SERVICE_NAME)
-        nginx_not_active = [
-            service for service in nginx_service.values() if not service.is_running()
-        ]
-        if not nginx_service or nginx_not_active:
+
+        if not self._is_service_running(container, synapse.SYNAPSE_NGINX_SERVICE_NAME):
             self.unit.status = ops.MaintenanceStatus("Waiting for NGINX")
             return
-        # All checks passed, the unit is active
-        self.model.unit.status = ops.ActiveStatus()
+
+        self.unit.status = ops.ActiveStatus()
+
+    def _is_service_running(self, container: ops.Container, service_name: str) -> bool:
+        """Check if all instances of a service are running.
+
+        Returns:
+            True if all services are running.
+        """
+        services = container.get_services(service_name)
+        return bool(services) and all(service.is_running() for service in services.values())
 
     def _on_register_user_action(self, event: ActionEvent) -> None:
         """Register user and report action result.
