+++ tests/unit_harness/test_backup_observer.py
