+++ tests/unit_harness/test_exceptions.py
