+++ tests/unit_harness/test_media_observer.py
