+++ tests/unit_harness/test_register_user_action.py
