+++ tests/unit_harness/test_synapse_workload.py
